requirements.txt
streamlit
groq
python-dotenv
elevenlabs
SpeechRecognition
pydub
streamlit-mic-recorder